+++ oraclepaas/config.go
@@ -8,26 +8,32 @@ import (
 	"strings"
 
 	"github.com/hashicorp/go-cleanhttp"
+	"github.com/hashicorp/go-oracle-terraform/application"
 	"github.com/hashicorp/go-oracle-terraform/database"
 	"github.com/hashicorp/go-oracle-terraform/java"
+	"github.com/hashicorp/go-oracle-terraform/mysql"
 	"github.com/hashicorp/go-oracle-terraform/opc"
 	"github.com/hashicorp/terraform/helper/logging"
 	"github.com/hashicorp/terraform/terraform"
 )
 
 type Config struct {
-	User             string
-	Password         string
-	IdentityDomain   string
-	MaxRetries       int
-	Insecure         bool
-	DatabaseEndpoint string
-	JavaEndpoint     string
+	User                string
+	Password            string
+	IdentityDomain      string
+	MaxRetries          int
+	Insecure            bool
+	DatabaseEndpoint    string
+	JavaEndpoint        string
+	ApplicationEndpoint string
+	MySQLEndpoint       string
 }
 
 type OPAASClient struct {
-	databaseClient *database.Client
-	javaClient     *java.Client
+	databaseClient    *database.Client
+	javaClient        *java.Client
+	applicationClient *application.Client
+	mysqlClient       *mysql.MySQLClient
 }
 
 func (c *Config) Client() (*OPAASClient, error) {
@@ -85,6 +91,32 @@ func (c *Config) Client() (*OPAASClient, error) {
 			return nil, err
 		}
 		oraclepaasClient.javaClient = javaClient
+	}
+
+	if c.ApplicationEndpoint != "" {
+		applicationEndpoint, err := url.ParseRequestURI(c.ApplicationEndpoint)
+		if err != nil {
+			return nil, fmt.Errorf("Invalid application endpoint URI: %+v", err)
+		}
+		config.APIEndpoint = applicationEndpoint
+		applicationClient, err := application.NewClient(&config)
+		if err != nil {
+			return nil, err
+		}
+		oraclepaasClient.applicationClient = applicationClient
+	}
+
+	if c.MySQLEndpoint != "" {
+		mysqlEndpoint, err := url.ParseRequestURI(c.MySQLEndpoint)
+		if err != nil {
+			return nil, fmt.Errorf("Invalid jmysqlava endpoint URI: %+v", err)
+		}
+		config.APIEndpoint = mysqlEndpoint
+		mysqlClient, err := mysql.NewMySQLClient(&config)
+		if err != nil {
+			return nil, err
+		}
+		oraclepaasClient.mysqlClient = mysqlClient
 	}
 
 	return oraclepaasClient, nil